refactor(settings): move reader card into card folder
